Extract speaker into a reusable package
