+++ internal/speaker/layer2_controller_test.go
@@ -1,4 +1,4 @@
-package main
+package speaker
 
 import (
 	"net"
